Déplace le code relatif à la licence dans un fichier dédié
