+++ zds/tutorialv2/urls/urls_contents.py
@@ -1,7 +1,8 @@
 from django.urls import path
 from django.views.generic.base import RedirectView
 
-from zds.tutorialv2.views.contents import CreateContent, EditContent, EditContentLicense, DeleteContent
+from zds.tutorialv2.views.contents import CreateContent, EditContent, DeleteContent
+from zds.tutorialv2.views.licence import EditContentLicense
 from zds.tutorialv2.views.display.container import ContainerValidationView
 from zds.tutorialv2.views.display.content import ContentValidationView
 from zds.tutorialv2.views.events import EventsList
